+++ src/main/java/gift/user/dto/UserSingupRequestDto.java
@@ -1,0 +1,14 @@
+package gift.user.dto;
+
+public class UserSingupRequestDto {
+    private String email;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
